abort this approach
reason: using .graphql file to generate a useable schema is not a very good approach. because it is quite difficult to implement Connection interface (for the global id part? possibly we still can do it.)

possible solution: we can mannually write resolver to resolve those problems.

HOWEVER! the most critical reason to abandon writing .graphql file is reusability!
how can you reuse a .graphql file? a much better way is to write .js file and extract concept on it and build reusable components.

so this comes an end to this branch. I will probably still update it in future.